% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{about.soda}
\alias{about.soda}
\title{Return rich metadata about an asset in a list}
\usage{
\method{about}{soda}(request, ...)
}
\arguments{
\item{request}{a SODA request}

\item{...}{dot-dot-dot}
}
\description{
Return rich metadata about an asset in a list
}
