% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{as_data_frame.soda}
\alias{as_data_frame.soda}
\title{Retrieve and parse a SODA response as data frame}
\usage{
\method{as_data_frame}{soda}(x, guess_type = TRUE, ...)
}
\arguments{
\item{x}{a SODA request}

\item{...}{dot-dot-dot, ignored}
}
\description{
Retrieve and parse a SODA response as data frame
}
