% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{catalog}
\alias{catalog}
\title{Metadata on datasets on a Socrata domain}
\usage{
catalog(domain, rich = FALSE)
}
\arguments{
\item{domain}{a Socrata domain}

\item{rich}{get rich metadata}
}
\value{
a data frame
}
\description{
Metadata on datasets on a Socrata domain
}
