% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{collect.soda}
\alias{collect.soda}
\title{Submit a SODA query and retrieve result}
\usage{
\method{collect}{soda}(x, guess_parse = TRUE, ...)
}
\arguments{
\item{x}{a SODA request}

\item{guess_parse}{logical, use `readr::guess_parse()` on columns?}

\item{...}{dot-dot-dot, ignored}
}
\description{
Submit a SODA query and retrieve result
}
