% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{colnames.soda}
\alias{colnames.soda}
\title{Column names}
\usage{
\method{colnames}{soda}(request, ...)
}
\arguments{
\item{request}{a SODA request}

\item{...}{dot-dot-dot}
}
\description{
Column names
}
