% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{credential}
\alias{credential}
\title{Add credentials to a SODA request}
\usage{
credential(request, app_token = NULL, user = NULL, password = NULL,
  ...)
}
\arguments{
\item{request}{a SODA request}

\item{app_token}{an application token issued by Socrata}

\item{user}{use name issued by the Socrata domain}

\item{password}{password useed by the Socrata domain}

\item{...}{dot-dot-dot, ignored}
}
\value{
a SODA request
}
\description{
Add credentials to a SODA request
}
