% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{filter.soda}
\alias{filter.soda}
\title{Return rows with matching conditions}
\usage{
\method{filter}{soda}(.data, ...)
}
\arguments{
\item{.data}{a SODA request}

\item{...}{dot-dot-dot}
}
\description{
Return rows with matching conditions
}
