% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{glimpse.soda}
\alias{glimpse.soda}
\title{Take a glimpse of the first of an asset}
\usage{
\method{glimpse}{soda}(request, n = 20)
}
\arguments{
\item{request}{a SODA request}

\item{n}{number of rows to query}
}
\description{
Take a glimpse of the first of an asset
}
