% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{head.soda}
\alias{head.soda}
\title{Return the first part of an asset}
\usage{
\method{head}{soda}(x, n = 10, ...)
}
\arguments{
\item{x}{a SODA request}

\item{n}{number of rows to query}

\item{...}{dot-dot-dot}
}
\description{
Queries the first part of an asset using a SODA request
Returns 10 rows by default
}
