% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{print.soda}
\alias{print.soda}
\title{Print brief information about an asset targeted by a SODA request to the screen}
\usage{
\method{print}{soda}(x, ...)
}
\arguments{
\item{x}{a SODA request}

\item{...}{dot-dot-dot}
}
\description{
Print brief information about an asset targeted by a SODA request to the screen
}
