% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{query.soda}
\alias{query.soda}
\title{Use a plain SoQL query string in a SODA request}
\usage{
\method{query}{soda}(request, query_string, ...)
}
\arguments{
\item{request}{a SODA request}

\item{query_string}{a SoQL query string}

\item{...}{dot-dot-dot, ignored}
}
\value{
a SODA request with query
}
\description{
Forgiving parameterized queries built by tidyish verbs
}
