% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{response.soda}
\alias{response.soda}
\title{Return response from a SODA request}
\usage{
\method{response}{soda}(request, ...)
}
\arguments{
\item{request}{a SODA request}

\item{...}{dot-dot-dot, ignore}
}
\description{
Return response from a SODA request
}
