% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{response_content}
\alias{response_content}
\title{Get content from the response to a request}
\usage{
response_content(...)
}
\arguments{
\item{...}{dot-dot-dot}
}
\description{
This is a generic function
}
