% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{response_content.soda}
\alias{response_content.soda}
\title{Return content of response from a SODA request}
\usage{
\method{response_content}{soda}(request, ...)
}
\arguments{
\item{request}{a SODA request}

\item{...}{dot-dot-dot, ignored}
}
\description{
Return content of response from a SODA request
}
