% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{show_query.soda}
\alias{show_query.soda}
\title{Returns the query of a SODA request}
\usage{
\method{show_query}{soda}(x, ...)
}
\arguments{
\item{x}{a SODA request}

\item{...}{dot-dot-dot, ignored}
}
\description{
Returns the query of a SODA request
}
