% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{soda}
\alias{soda}
\title{Create a SODA request}
\usage{
soda(endpoint, domain, uuid, ...)
}
\arguments{
\item{endpoint}{a url pointing to a SODA endpoint.}

\item{domain}{a Socratas domain}

\item{uuid}{the "four-by-four" identifier of an asset on the domain}

\item{...}{dot-dot-dot}
}
\value{
a list containing the RESTful API request. See httr::parse_url.
}
\description{
Create a SODA request to an asset either using a url,
or using together the domain name and the identifier.
}
