% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{soda_data_types}
\alias{soda_data_types}
\title{Get SODA data types of an an asset}
\usage{
soda_data_types(request, ...)
}
\arguments{
\item{request}{a SODA request}

\item{...}{dot-dot-dot, ignored}
}
\description{
Get SODA data types of an an asset
}
