% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{summarise.soda}
\alias{summarise.soda}
\title{Reduce multiple values down to a single value}
\usage{
\method{summarise}{soda}(.data, ...)
}
\arguments{
\item{.data}{a SODA request}

\item{...}{dot-dot-dot}
}
\description{
Reduce multiple values down to a single value
}
