% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{write_csv.soda}
\alias{write_csv.soda}
\title{Write a SODA query return to CSV file}
\usage{
\method{write_csv}{soda}(x, path, na = "NA")
}
\arguments{
\item{x}{a SODA query}

\item{path}{path in file system to write to}

\item{na}{String used for missing values}
}
\description{
Write a SODA query return to CSV file
}
