% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{collect.soda}
\alias{collect.soda}
\title{Submit a SODA query and retrieve result}
\usage{
\method{collect}{soda}(x, guess_type = TRUE, ...)
}
\arguments{
\item{x}{a SODA request}

\item{guess_type}{logical, guess data types?}

\item{...}{dot-dot-dot, ignored}
}
\description{
Submit a SODA query and retrieve result
}
